remount on theme change
